refactor inspector into mixins and improve storing
