+++ earningspy/common/constants.py
@@ -33,3 +33,4 @@ EARNING_VIX_KEY='EARNING_VIX'
 CAR_KEY='CAR_{}'
 BHAR_KEY='BHAR_{}'
 DATADATE_KEY='DATADATE'
+ALLOWED_WINDOWS = [3, 30, 60]
